// SDKs/twilio.com/twilio_accounts_v1/1.36.0/typescript/src/sdk/models/shared/accountsv1credentialcredentialaws.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class AccountsV1CredentialCredentialAws extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}